+++ examples/user.py
@@ -52,29 +52,19 @@ oneview_client = OneViewClient(config)
 users = oneview_client.users
 
 # Create a User
-#user = users.create(options)
-#print("Created user '%s' successfully.\n  uri = '%s'\n" % (user['userName'], user['uri']))
-#
-## Change Password
-#change_password_request = {
-#   "currentPassword": "admin12345",
-#   "enabled": "true",
-#   "password": "admin1234",
-#   "userName": "venki"
-#}
-#changePasswordResponse = users.change_password(change_password_request)
-#print("Changed Password successfully")
-#print(changePasswordResponse)
+user = users.create(options)
+print("Created user '%s' successfully.\n  uri = '%s'\n" % (user['userName'], user['uri']))
 
-# Adds different roles to the recently created user
-#user = oneview_client.users.get_by_userName(options['userName'])
-#print(user)
-#print("\n user: %s, roles before update: %s" % (user['userName'], user['roles']))
-#user_copy = deepcopy(user)
-#user_copy['replaceRoles'] = True
-#user_copy['roles'] = ['Infrastructure administrator']
-#user_update = user.update(user_copy)
-#print("\n user: %s, roles after update: %s" % (user_update['userName'], user_update['roles']))
+# Change Password
+change_password_request = {
+  "currentPassword": "admin12345",
+  "enabled": "true",
+  "password": "admin1234",
+  "userName": "admin"
+}
+changePasswordResponse = users.change_password(change_password_request)
+print("Changed Password successfully")
+print(changePasswordResponse)
 
 # Get user by role
 input_role = 'Infrastructure administrator'
@@ -97,7 +87,3 @@ print("Is full name already in use? %s" % (bol))
 # # Validates if user name is already in use
 bol = oneview_client.users.validate_user_name(options['userName'])
 print("Is user name already in use? %s" % (bol))
-
-# Delete the created user
-user.delete()
-print("\nSuccessfully deleted user")