+++ tests/unit/resources/security/test_users.py
@@ -21,7 +21,7 @@ import mock
 
 from hpeOneView.connection import connection
 from hpeOneView.resources.security.users import Users
-from hpeOneView.resources.resource import ResourceClient
+from hpeOneView.resources.resource import Resource, ResourceHelper
 from hpeOneView.exceptions import HPEOneViewException
 
 
@@ -31,7 +31,7 @@ class UsersTest(unittest.TestCase):
         self.connection = connection(self.host, 800)
         self._users = Users(self.connection)
 
-    @mock.patch.object(ResourceClient, 'get_all')
+    @mock.patch.object(Resource, 'get_all')
     def test_get_all_called_once(self, mock_get_all):
         filter = 'name=TestName'
         sort = 'name:ascending'
@@ -40,7 +40,7 @@ class UsersTest(unittest.TestCase):
 
         mock_get_all.assert_called_once_with(2, 500, filter=filter, sort=sort)
 
-    @mock.patch.object(ResourceClient, 'create')
+    @mock.patch.object(Resource, 'create')
     def test_create_should_use_given_values(self, mock_create):
         resource = {
             'enabled': 'true',
@@ -59,63 +59,42 @@ class UsersTest(unittest.TestCase):
         mock_create.assert_called_once_with(resource_rest_call, timeout=30,
                                             default_values=self._users.DEFAULT_VALUES)
 
-    @mock.patch.object(ResourceClient, 'update')
-    def test_update_should_use_given_values(self, mock_update):
-        resource = {
-            'enabled': 'true',
-            'fullName': 'testUser101',
-            'mobilePhone': '555-2121',
-            'officePhone': '555-1212',
-            'password': 'myPass1234',
-            'roles': ['Read only'],
-            'type': 'UserAndRoles',
-            'userName': 'testUser'
-        }
-        resource_rest_call = resource.copy()
-        mock_update.return_value = {}
-
-        self._users.update(resource, 60)
-        mock_update.assert_called_once_with(resource_rest_call, timeout=60,
-                                            default_values=self._users.DEFAULT_VALUES, uri='/rest/users')
-
-    @mock.patch.object(ResourceClient, 'delete')
-    def test_delete_called_once(self, mock_delete):
-        id = 'fake'
-        self._users.delete(id, force=False, timeout=-1)
-
-        mock_delete.assert_called_once_with(id, force=False, timeout=-1)
-
-    @mock.patch.object(ResourceClient, 'get')
-    def test_get_by_called_once(self, mock_get):
-        self._users.get_by('userName', 'OneViewSDK Test User')
+    @mock.patch.object(Resource, 'do_get')
+    def test_get_by_called_with_userName(self, mock_get):
+        self._users.get_by_name('OneViewSDK Test User')
         mock_get.assert_called_once_with('/rest/users/OneViewSDK Test User')
 
-    @mock.patch.object(ResourceClient, 'get')
+    @mock.patch.object(Resource, 'do_get')
     def test_get_by_called_with_role(self, mock_get):
-        self._users.get_by('role', 'fakerole')
+        self._users.get_by_role('fakerole')
         mock_get.assert_called_once_with('/rest/users/roles/users/fakerole')
 
-    @mock.patch.object(ResourceClient, 'get')
-    def test_get_by_called_with_something_invalid(self, mock_get):
-        try:
-            self._users.get_by('test', 'test')
-        except HPEOneViewException as exception:
-            self.assertEqual('Only userName, name and role can be queried for this resource.', exception.args[0])
-        else:
-            self.fail("Expected Exception was not raised")
-
-    @mock.patch.object(ResourceClient, 'create_with_zero_body')
-    def test_validate_full_name_called_once(self, mock_create_with_zero_body):
+    @mock.patch.object(Resource, 'do_post')
+    def test_validate_full_name_called_once(self, mock_post):
 
         self._users.validate_full_name('fullname101')
 
         expected_uri = '/rest/users/validateUserName/fullname101'
-        mock_create_with_zero_body.assert_called_once_with(uri=expected_uri, timeout=-1)
+        mock_post.assert_called_once_with(expected_uri, {}, -1, None)
 
-    @mock.patch.object(ResourceClient, 'create_with_zero_body')
-    def test_validate_user_name_called_once(self, mock_create_with_zero_body):
+    @mock.patch.object(Resource, 'do_post')
+    def test_validate_user_name_called_once(self, mock_post):
 
         self._users.validate_user_name('userName')
 
         expected_uri = '/rest/users/validateLoginName/userName'
-        mock_create_with_zero_body.assert_called_once_with(uri=expected_uri, timeout=-1)
+        mock_post.assert_called_once_with(expected_uri, {}, -1, None)
+
+    @mock.patch.object(Resource, 'ensure_resource_data')
+    @mock.patch.object(ResourceHelper, 'do_put')
+    def test_change_password(self, mock_put, mock_ensure_resource_data):
+        request = {
+            "currentPassword": "admin12345",
+            "enabled": "true",
+            "password": "admin1234",
+            "userName": "admin"
+        }           
+        self._users.change_password(request)
+
+        expected_uri = '/rest/users/'
+        mock_put.assert_called_once_with(expected_uri, request, -1, None)