+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Refactored base class to take default API version as per provided Oneview applia
 
 #### Bug fixes & Enhancements
 - [#81](https://github.com/HewlettPackard/oneview-python/issues/81) EthernetNetworks Update does not work.
+- [#91](https://github.com/HewlettPackard/oneview-python/issues/91) Change Password Sample?
 
 # 5.3.0
 #### Notes